Types have their own file. Module additions.
